requirements.txt
Flask==0.10
Flask-Script==2.0.5
Flask-Cache==0.12
Flask-Mail==0.9.1
Flask-Cors==1.10.2
Flask-Stormpath==0.4.1
Flask-SQLAlchemy==2.0
SQLAlchemy==0.9.2
blinker==1.3
alembic==0.6.3
redis==2.9.1
gunicorn==0.17.2
python-dateutil==2.4.0
PyYAML==3.10
psycopg2==2.4.6